server/project/requirements.txt
Django==1.9.1
pyserial==3.0.1
